Break up  file into separate files, refactor  method to be utility function, and rename  to
