+++ src/biochemical_data_connectors/__init__.py
@@ -2,15 +2,12 @@
 BiochemicalDataConnectors: A Python package to extract chemical
 and biochemical from public databases.
 """
-from biochemical_data_connectors.connectors.bioactive_compounds_connectors import (
-    BaseBioactivesConnector,
-    ChEMBLBioactivesConnector,
-    PubChemBioactivesConnector
-)
-
+from biochemical_data_connectors.connectors.bioactive_compounds.base_bioactives_connector import BaseBioactivesConnector
+from biochemical_data_connectors.connectors.bioactive_compounds.chembl_bioactives_connector import ChEMBLBioactivesConnector
+from biochemical_data_connectors.connectors.bioactive_compounds.pubchem_bioactives_connector import PubChemBioactivesConnector
 from biochemical_data_connectors.connectors.ord_connectors import OpenReactionDatabaseConnector
-
 from biochemical_data_connectors.utils.api.mappings import uniprot_to_gene_id_mapping, pdb_to_uniprot_id_mapping
+from biochemical_data_connectors.utils.standardization_utils import CompoundStandardizer
 
 __all__ = [
     # --- Base Classes ---
@@ -24,4 +21,5 @@ __all__ = [
     # --- Public Utility Functions ---
     "uniprot_to_gene_id_mapping",
     "pdb_to_uniprot_id_mapping",
+    "CompoundStandardizer"
 ]
